@imkero 2025-05-05
To make the benchmark reproducible across different runs, I suggest setting `--input-token-distribution uniform 200 200` and `--output-token-distribution uniform 200 200`, and also pass image data URLs instead of image http URLs