@imkero 2025-05-05
> > I feel that those changes should not take too much effect on the E2E time cost as they cost <0.1ms each op... even x1000 will bring only 0.1s difference
> 
> ms or us? 0.1 ms * 1000 = 100s

sorry, i mean 0.1ms * 1000 = 0.1s in total