@vadiklyutiy 2025-05-05
> and also pass image data URLs instead of **downloading image via http URLs**

Behaviour with the same image for 1000 requests and different images for 1000 requests is a bit different. So, different images was intentionally implemented in diff. 
In perform results above I made 21 runs for before and after and took median. Also checked with nsys. `_prepare_inputs` takes around 5% before and around 3% after what pretty good corresponding E2E results.