@imkero 2025-05-05
Alright i forgot multiplying the output tokens num.

Just benchmarking piecewise for 200,000 times `get_next_input_positions_tensor` takes nearly 1sec so it does make sense i think