@vadiklyutiy 2025-05-05
> , I suggest setting `--input-token-distribution uniform 200 200` and `--output-token-distribution uniform 200 200`

Agree, this make results more stable