@vadiklyutiy 2025-05-05
> And maybe please take a look on #16881 ? It optimize the rest part (new request) of MRoPE position preparation, and i will take `get_next_input_positions_tensor` into consideration because they are reported to be time-costing as well.
> 
> Thanks @vadiklyutiy for reporting this with benchmark

If numpy+numba speed ups `get_next_input_positions_tensor` by >10x we can definitely waive caching. By my intuition don't provide any estimation what speed up will we get :) 
If you give some numbers for speed up of your implementation of `get_next_input_positions_tensor` it would great.