@imkero 2025-05-05
> > I have confirmed some speedup with this PR's changes
> > > Tested on NVIDIA A10
> > 
> > 
> > ```
> > vllm serve Qwen/Qwen2.5-VL-3B-Instruct --disable-log-requests --max-num-seqs 128 --block-size 16 --max-num-batched-tokens 8192 --max-model-len 8192 --no-enable-prefix-caching
> > ```
> > 
> > 
> >     
> >       
> >     
> > 
> >       
> >     
> > 
> >     
> >   
> > ```
> > fib benchmark -rps 50 --input-token-distribution uniform 200 200 --output-token-distribution uniform 200 200 --num-of-imgs-per-req 1 --img-ratios-per-req 512x512 -n 1024 --base-url http://localhost:8000 --endpoint v1/chat/completions --backend openai-chat
> > ```
> > 
> > 
> >     
> >       
> >     
> > 
> >       
> >     
> > 
> >     
> >   
> > code base	duration(s)	req/s	input tokens/s	output tokens/s
> > main branch	172.86	5.92	3231.29	1184.74
> > PR #17617	171.74	5.96	3252.25	1192.53
> 
> I tested with H100. reqs/s for H100 is around 30. So, with A10 model itself takes longer. `_prepare_input` including `get_next_input_positions_tensor` is CPU part and time of `_prepare_input` doesn't depend what GPU we have. So, the less improvement(in %) with A10 fit with the theory.

Yes, I think for CPU overhead focusing on the `duration` section would make sense (-1.1 secs)