@vadiklyutiy 2025-05-05
> And the result is here:
> 
> ```
> run_torch: 1.990s
> run_numba: 0.066s
> run_torch_lru: 0.872s
> ```

Results look awesome! @imkero Thank you for adding `get_next_input_positions_tensor` in #16881. Waiting of merge of  #16881