@imkero 2025-05-05
> Results look awesome! @imkero Thank you for adding `get_next_input_positions_tensor` in #16881. Waiting of merge of #16881

Pretty thanks for pointing out the time comsumption of `get_next_input_positions_tensor`.